Setting up Basic Behavior for the Waves
Created a Quad Plane to be deformed by the wave, its material that will respond to said deformation and render. Added the BASE C# Script for the Basic Behavior of the Discretisized Differencial Equation of the Wave Phenomena.
Script still Requires Debugging, Object Quad needs to be tied to the Script and the Script need to call the Object Mesh to Deform it.