Grid Mapped Wave & Colored Mesh Mapped Wave
The Grid map is slow and grows in complexity very fast, but shows well the behavior of a wave that propagates indefinetly (no loss of energy and a presence of tangible boundaries) and will refflect in the grid boundaries, it also shows how bending effects can create a new wave pattern with a different direction of propagation.
The Mesh Mapped Wave Need fixes to behave like the grid one while keeping the optmization of using a mesh.
The camera cull should be fixed to show all the simulations being done, and the boundary sizes should be the same for all simulations.